+++ docs/usage.md
@@ -1,0 +1,5 @@
+---
+sidebar_label: Usage
+---
+
+# Usage
