+++ package.json
@@ -33,18 +33,12 @@
   ],
   "homepage": "https://skelosaurus.com",
   "dependencies": {
-    "colors": "^1.4.0",
     "commander": "^12.0.0",
     "file-easy": "^1.0.3",
-    "file-set": "^4.0.1",
-    "fs": "^0.0.1-security",
     "glob": "^10.3.10",
     "hbsr": "^1.0.11",
     "jest": "^29.7.0",
     "lodash": "^4.17.21",
-    "lorem-ipsum": "^2.0.3",
-    "markdown": "^0.5.0",
-    "path": "^0.12.7",
     "shelljs": "^0.8.5",
     "yamljs": "^0.3.0"
   },
